feat: parametrize e2e tests across dev server and static site
Rename test app package from app to my_app to exercise underscore-
containing package names in wheel filename normalization. Add
serving_mode parametrized fixture so all e2e tests run against both
dev server and generated static site. Add console_errors and
assert_no_python_errors fixtures to detect Python tracebacks in the
browser console, catching micropip/wheel parse failures.
