+++ tests/e2e/my_app/pages/lifecycle.py
