+++ tests/e2e/my_app/pages/switch_test.py
